requirements.txt
Flask==0.10.1
python-ldap==2.4.22
uWSGI==2.0.11.2
